Add IProduct interface and Product class for data access
A new `IProduct` interface was introduced in the `Repository.Interface`
namespace, defining a `GetProducts()` method to retrieve product data.

The `Product` class in the `Repository` namespace implements this
interface. It connects to a SQL Server database, executes the
`sp_GetAllProducts` stored procedure, maps the results to a list of
`Common.Product` objects using `SqlDataReader`, and returns the list.

Namespaces used include `Microsoft.Data.SqlClient`, `System.Data`, and
`Repository.Interface`.